Move user filters tests into a standalone module
